# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Hôtel Artichaut",
        "version": "1.0"
    },
    "paths": {
...
        "api/reservation/{id}": {
            "get": {
                "tags": [
                    "Réservations"
                ],
                "summary": "Afficher une réseravtion par ID",
                "operationId": "3f1635e080531a12db03f217ab9defeb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la réseravtion",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail de la réservation"
                    }
                }
            },
            "put": {
                "tags": [
                    "Réservations"
                ],
                "summary": "Modifier une réservation existant",
                "operationId": "d7040c82cb732d9b7eca75e922ed76b2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la réservation",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "startDate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025/05/01"
                                    },
                                    "endDate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025/05/03"
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": "254"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Réservation mis à jour"
                    }
                }
            },
            "post": {
                "tags": [
                    "Réservations"
                ],
                "summary": "Modifier une réservation existant",
                "operationId": "2e057d8c210ab30b94001c699581c666",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la réservation",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "startDate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025/05/01"
                                    },
                                    "endDate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025/05/03"
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": "254"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Réservation mis à jour"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Réservations"
                ],
                "summary": "Supprimer une réservation",
                "operationId": "369de45e8ffaf63600544046591afec2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la réseravtion à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Réservation supprimé"
                    }
                }
            }
        },
...
    },
...
}